requirements.txt
Django
setuptools
Pillow
South
Unipath
argparse
dateutils
distribute
django-announcements
django-appconf
django-compressor
django-filebrowser
django-grappelli
django-guardian
django-indexer
django-paging
django-report-builder
django-taggit==0.10
django-templatetag-sugar
django-userena
django-widget-tweaks
djangorestframework
docutils
easy-thumbnails
eventlet
feedparser
greenlet
lockfile
mimeparse
openpyxl
psycopg2
python-daemon
python-dateutil
pytz
raven
simplejson
six
wsgiref

#IMPORTANT:
#pip install --upgrade setuptools
# export ARCHFLAGS="-Wno-error=unused-command-line-argument-hard-error-in-future"